board: ti: am57xx: Switch HDMI muxing on latest X15/GPEVM boards
HDMI_LS_OE from GPIO6_28(Y9) to GPIO2_30 (AG8) to avoid to avoid a
1.8V GPIO toggling a 3.3V input when the SD card in HS 1.8V mode. This
requires a minor change in our configuration for latest X15/GPEVM setup.
We split the HDMI LS_OE configuration based on the board revisions
and by default configure the pinmux for MMC1_SDWP as driver_off.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>