drivers: mmc: omap_hsmmc: add support to set default io voltage
"ti,dual-volt" is used in linux kernel to set the voltage capabilities.
For host controller dt nodes that doesn't have "ti,dual-volt",
it's assumed 1.8v is the io voltage. This is not always true (like in
the case of beagle-x15 where the io lines are connected to 3.3v).
So add a new optional dt property _iov_ to set the default io voltage.

If "iov" property is not present, the default io voltage of 1.8v will
be used. However if "ti,dual-volt" is present, the default io voltage
will be set to 3v

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>