ARM: DRA7: Revise DSP/IVA clock frequency update logic
The commit 539982820939 ("ARM: DRA7: Fixup DSPEVE and IVA clock
frequencies based on OPP") only updates the 'assigned-clock-rates'
property and expects that the 'assigned-clocks' property is already
defined in the existing DTBs. The kernel clock init configuration
logic requires both these properties to be defined in a clock
node inorder to properly update the clock frequencies. Enhance
the current frequency update logic to also add the 'assigned-clocks'
property so that the DSPEVE and IVA DPLLs are configured properly
even for kernels that do not have the OPP_NOM DPLL clock rate
configuration in the DTBs. The additional logic is a no-op for
kernels that do use DTBs with corresponding 'assigned-clocks'
property.

Signed-off-by: Suman Anna <s-anna@ti.com>