arm: omap: Introduce vcores_update function
This is needed in case of am57xx where in board detection is done
pretty late in the init sequence. The pmic registers for variants
of am57xx are different hence we need to assign them carefully based
on the board type. Add a function to assign omap_vcores after the
board detection is done.

Fixes: 888ea8611c1 ("ARM: dts: AM572x-IDK Initial Support")
Fixes: 1084ffa86a5 ("ARM: dts: AM571x-IDK Initial Support")
Reported-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>