board: ti: AM57xx: Setup AM571x-idk GPIO to glow on MLO start
When the board hangs while boot, it is hard to identify without a serial
port about which stage the board failed to boot at. So setup pinmux for
GPIO LED so that it glows by default, as HLOS comes around later on, it
can choose to drive the GPIO to override the state and override the weak
control module pull as it desires.

Signed-off-by: Nishanth Menon <nm@ti.com>