spl: omap: enable cache for all sections of SRAM
Enabling of cache for the last section of the SRAM was missed and
was not caught in testing as the spl size is less. With usb rndis
boot on AM437x, spl size reaches the last section and results in
a data abort. Fixing it by enabling the cache for the last
section as well.

Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>