ARM: DRA7: Fixup GPU clock frequencies based on OPP
This patch adds support to update the device-tree blob to adjust the
GPU DPLL clocks pertinent to the selected OPP choice, with
the default being OPP_NOM.

The GPU DPLL is not essential for u-boot functionality. The non-essential
DPLLs are configured within the kernel during the clock init step when
parsing the device tree and creating the clock devices. This approach
meets both the u-boot and kernel needs.

Based on logic similar to that of DSPEVE and IVA voltage domains in
commit 5399828 ("ARM: DRA7: Fixup DSPEVE and IVA clock frequencies based
on OPP")

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
Acked-by: Suman Anna <s-anna@ti.com>